adds deployment form component
