+++ dashboard/src/components/DeploymentForm/index.tsx
@@ -1,0 +1,3 @@
+import DeploymentForm from "./DeploymentForm";
+
+export default DeploymentForm;
